22-feat : 글 상세 컴포넌트 패키지 컴포넌트로 변경
